Padrão doc swagger
- Criado padrão de doc swagger com os DTO
- Criado padrão de retorno de response code e exceções

+++ backend/src/main/java/com/utfpr/donare/service/UserService.java
@@ -2,8 +2,11 @@ package com.utfpr.donare.service;
 
 import com.utfpr.donare.config.jwt.JwtTokenUtil;
 import com.utfpr.donare.domain.User;
-import com.utfpr.donare.dto.UserResponseDTO;
 import com.utfpr.donare.dto.UserRequestDTO;
+import com.utfpr.donare.dto.UserResponseDTO;
+import com.utfpr.donare.exception.BadRequestException;
+import com.utfpr.donare.exception.ResourceNotFoundException;
+import com.utfpr.donare.exception.UnauthorizedException;
 import com.utfpr.donare.mapper.UserMapper;
 import com.utfpr.donare.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
@@ -15,6 +18,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
@@ -28,6 +32,17 @@ public class UserService implements UserDetailsService {
     @Transactional
     public UserResponseDTO save(UserRequestDTO dto) {
 
+        if (userRepository.findByEmail(dto.getEmail()).isPresent()) {
+
+            throw new BadRequestException("O e-mail '" + dto.getEmail() + "' já está em uso.");
+        }
+
+        if (userRepository.findByCpfOuCnpj(dto.getCpfOuCnpj()).isPresent()) {
+
+            throw new BadRequestException("O CPF/CNPJ '" + dto.getCpfOuCnpj() + "' já está cadastrado.");
+        }
+
+        // todo verficar para usar mapper com o encoder!
         User user = User.builder()
                 .nome(dto.getNome())
                 .email(dto.getEmail())
@@ -44,13 +59,19 @@ public class UserService implements UserDetailsService {
 
     @Transactional
     public void delete(Long id) {
+
+        if (!userRepository.existsById(id)) {
+
+            throw new ResourceNotFoundException("Usuário com ID " + id + " não encontrado para exclusão.");
+        }
+
         userRepository.deleteById(id);
     }
 
     @Transactional
     public void update(Long id, UserRequestDTO dto) {
 
-        User user = userRepository.findById(id).orElseThrow(() -> new RuntimeException("Usuário não encontrado"));
+        User user = userRepository.findById(id).orElseThrow(() -> new BadRequestException("Usuário não encontrado"));
         user.setNome(dto.getNome());
         user.setEmail(dto.getEmail());
         user.setCpfOuCnpj(dto.getCpfOuCnpj());
@@ -68,19 +89,23 @@ public class UserService implements UserDetailsService {
         User user = findByEmail(email);
 
         if (!passwordEncoder.matches(senha, user.getPassword())) {
-            throw new RuntimeException("Credenciais inválidas");
+
+            throw new UnauthorizedException("Credenciais inválidas. E-mail ou senha incorretos.");
         }
 
         return jwtTokenUtil.autenticar(user);
     }
 
     public User findByEmail(String email) {
-        return userRepository.findByEmail(email)
-                .orElseThrow(() -> new RuntimeException("Usuário não encontrado"));
+
+        return userRepository.findByEmail(email).orElseThrow(() -> new ResourceNotFoundException("Usuário não encontrado com e-mail: " + email));
     }
 
-    public List<User> findAll() {
-        return userRepository.findAll();
+    public List<UserResponseDTO> findAllUsersDTO() {
+
+        return userRepository.findAll().stream()
+                .map(userMapper::toUserResponseDTO)
+                .collect(Collectors.toList());
     }
 
     public String compartilharCampanha(Long idCampanha) {
@@ -95,7 +120,7 @@ public class UserService implements UserDetailsService {
 
     @Override
     public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
-        return userRepository.findByEmail(email)
-                .orElseThrow(() -> new UsernameNotFoundException("Usuário não encontrado com email: " + email));
+
+        return userRepository.findByEmail(email).orElseThrow(() -> new UsernameNotFoundException("Usuário não encontrado com e-mail: " + email));
     }
 }
